Replace bool array with bitset for membership checking
Also use u8 to store digits rather than u32 so that boards can be
smaller and hopefully be more cache-efficient.

+++ examples/p96.rs
@@ -8,7 +8,9 @@ use sudoku::Grid;
 fn get_first_three_digits(grid: &Grid) -> u32 {
     let solution = grid.solve().unwrap();
     let solution_digits = solution.as_slice();
-    solution_digits[0] * 100 + solution_digits[1] * 10 + solution_digits[2]
+    u32::from(solution_digits[0]) * 100
+        + u32::from(solution_digits[1]) * 10
+        + u32::from(solution_digits[2])
 }
 
 fn main() {
